requirements_gpu.txt
bitstring==4.2.3
datasets==2.21.0
numpy
pandas==2.2.3
plotly==5.22.0
scipy==1.14.1
seaborn==0.13.2
tiktoken
torchinfo==1.8.0
transformers
pyspellchecker==0.8.3
wandb==0.18.3
onnx
beautifulsoup4==4.13.4
dragonmapper==0.3.0
huggingface_hub==0.30.2
jieba==0.42.1
konlpy==0.6.0
MeCab==0.996.11
pykakasi==2.0.0
Requests==2.32.3
rich==14.0.0
sentencepiece
spacy
tqdm==4.66.4
lm_eval==0.4.8
# optimizer related
adabelief-pytorch
adan-pytorch
torch_optimizer
tensorboard
